[App Search] Add roleHasScopedEngines helper + small `roles/` cleanup (#94038)
* Split out roles/ into separate files

- in preparation for new scoped engines helper

* Add new roleHasScopedEngines helper
