+++ x-pack/plugins/enterprise_search/public/applications/app_search/utils/role/get_role_abilities.test.ts
